requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Hyperparameter optimization
optuna>=3.0.0

# Logging and visualization
tensorboard>=2.13.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0

# Testing
pytest>=7.0.0

# Optional: For downloading model checkpoints
huggingface-hub>=0.16.0
